refactor: separate into auth module & user module
