+++ src/user/user.controller.ts
@@ -1,5 +1,5 @@
 import { Body, Controller, Get, Put, UseGuards } from '@nestjs/common';
-import { AuthService } from './auth.service';
+import { UserService } from './user.service';
 import { JwtAuthGuard } from '../common/guards/jwt-auth.guard';
 import { CurrentUser } from '../common/decorators/current-user.decorator';
 import { AuthenticatedUser } from '../common/interfaces/authenticated-user.interface';
@@ -7,12 +7,12 @@ import { UpdateUserRequestDto } from './dto/update-user-request.dto';
 
 @Controller('user')
 export class UserController {
-  constructor(private authService: AuthService) {}
+  constructor(private userService: UserService) {}
 
   @Get()
   @UseGuards(JwtAuthGuard)
   getCurrentUser(@CurrentUser() user: AuthenticatedUser) {
-    return this.authService.getCurrentUser(user.id);
+    return this.userService.getCurrentUser(user.id);
   }
 
   @Put()
@@ -21,6 +21,6 @@ export class UserController {
     @CurrentUser() user: AuthenticatedUser,
     @Body() updateUserRequest: UpdateUserRequestDto,
   ) {
-    return this.authService.updateUser(user.id, updateUserRequest.user);
+    return this.userService.updateUser(user.id, updateUserRequest.user);
   }
 }
